Staging: vt6656: removed custom UCHAR/USHORT/UINT/ULONG/ULONGLONG typedefs
Cleared all checkpatch warnings but 'do not add new typedefs' ones.

+++ drivers/staging/vt6656/dpc.c
@@ -90,7 +90,7 @@ void
 s_vProcessRxMACHeader (
       PSDevice pDevice,
       PBYTE pbyRxBufferAddr,
-      UINT cbPacketSize,
+      unsigned int cbPacketSize,
       BOOL bIsWEP,
       BOOL bExtIV,
      PUINT pcbHeadSize
@@ -107,8 +107,8 @@ static BOOL s_bAPModeRxCtl(
 static BOOL s_bAPModeRxData (
      PSDevice pDevice,
      struct sk_buff *skb,
-     UINT     FrameSize,
-     UINT     cbHeaderOffset,
+     unsigned int     FrameSize,
+     unsigned int     cbHeaderOffset,
      signed int      iSANodeIndex,
      signed int      iDANodeIndex
     );
@@ -117,7 +117,7 @@ static BOOL s_bAPModeRxData (
 static BOOL s_bHandleRxEncryption(
      PSDevice     pDevice,
      PBYTE        pbyFrame,
-     UINT         FrameSize,
+     unsigned int         FrameSize,
      PBYTE        pbyRsr,
      PBYTE       pbyNewRsr,
      PSKeyItem   * pKeyOut,
@@ -130,7 +130,7 @@ static BOOL s_bHostWepRxEncryption(
 
      PSDevice     pDevice,
      PBYTE        pbyFrame,
-     UINT         FrameSize,
+     unsigned int         FrameSize,
      PBYTE        pbyRsr,
      BOOL         bOnFly,
      PSKeyItem    pKey,
@@ -165,14 +165,14 @@ void
 s_vProcessRxMACHeader (
       PSDevice pDevice,
       PBYTE pbyRxBufferAddr,
-      UINT cbPacketSize,
+      unsigned int cbPacketSize,
       BOOL bIsWEP,
       BOOL bExtIV,
      PUINT pcbHeadSize
     )
 {
     PBYTE           pbyRxBuffer;
-    UINT            cbHeaderSize = 0;
+    unsigned int            cbHeaderSize = 0;
     PWORD           pwType;
     PS802_11Header  pMACHeader;
     int             ii;
@@ -267,7 +267,7 @@ s_vGetDASA (
      PSEthernetHeader psEthHeader
     )
 {
-	UINT            cbHeaderSize = 0;
+	unsigned int            cbHeaderSize = 0;
 	PS802_11Header  pMACHeader;
 	int             ii;
 
@@ -319,7 +319,7 @@ BOOL
 RXbBulkInProcessData (
      PSDevice         pDevice,
      PRCB             pRCB,
-     ULONG            BytesToIndicate
+     unsigned long            BytesToIndicate
     )
 {
 
@@ -334,20 +334,20 @@ RXbBulkInProcessData (
     PQWORD          pqwTSFTime;
     PBYTE           pbyFrame;
     BOOL            bDeFragRx = FALSE;
-    UINT            cbHeaderOffset;
-    UINT            FrameSize;
+    unsigned int            cbHeaderOffset;
+    unsigned int            FrameSize;
     WORD            wEtherType = 0;
     signed int             iSANodeIndex = -1;
     signed int             iDANodeIndex = -1;
-    UINT            ii;
-    UINT            cbIVOffset;
+    unsigned int            ii;
+    unsigned int            cbIVOffset;
     PBYTE           pbyRxSts;
     PBYTE           pbyRxRate;
     PBYTE           pbySQ;
 #ifdef Calcu_LinkQual
     PBYTE           pby3SQ;
 #endif
-    UINT            cbHeaderSize;
+    unsigned int            cbHeaderSize;
     PSKeyItem       pKey = NULL;
     WORD            wRxTSC15_0 = 0;
     DWORD           dwRxTSC47_16 = 0;
@@ -373,7 +373,7 @@ RXbBulkInProcessData (
 
     //[31:16]RcvByteCount ( not include 4-byte Status )
     dwWbkStatus =  *( (PDWORD)(skb->data) );
-    FrameSize = (UINT)(dwWbkStatus >> 16);
+    FrameSize = (unsigned int)(dwWbkStatus >> 16);
     FrameSize += 4;
 
     if (BytesToIndicate != FrameSize) {
@@ -935,9 +935,9 @@ RXbBulkInProcessData (
         if (bIsWEP) {
             WORD        wLocalTSC15_0 = 0;
             DWORD       dwLocalTSC47_16 = 0;
-            ULONGLONG       RSC = 0;
+	    unsigned long long       RSC = 0;
             // endian issues
-            RSC = *((ULONGLONG *) &(pKey->KeyRSC));
+	    RSC = *((unsigned long long *) &(pKey->KeyRSC));
             wLocalTSC15_0 = (WORD) RSC;
             dwLocalTSC47_16 = (DWORD) (RSC>>16);
 
@@ -1152,7 +1152,7 @@ static BOOL s_bAPModeRxCtl (
 static BOOL s_bHandleRxEncryption (
      PSDevice     pDevice,
      PBYTE        pbyFrame,
-     UINT         FrameSize,
+     unsigned int         FrameSize,
      PBYTE        pbyRsr,
      PBYTE       pbyNewRsr,
      PSKeyItem   * pKeyOut,
@@ -1161,7 +1161,7 @@ static BOOL s_bHandleRxEncryption (
      PDWORD      pdwRxTSC47_16
     )
 {
-    UINT            PayloadLen = FrameSize;
+    unsigned int            PayloadLen = FrameSize;
     PBYTE           pbyIV;
     BYTE            byKeyIdx;
     PSKeyItem       pKey = NULL;
@@ -1298,7 +1298,7 @@ static BOOL s_bHandleRxEncryption (
 static BOOL s_bHostWepRxEncryption (
      PSDevice     pDevice,
      PBYTE        pbyFrame,
-     UINT         FrameSize,
+     unsigned int         FrameSize,
      PBYTE        pbyRsr,
      BOOL         bOnFly,
      PSKeyItem    pKey,
@@ -1309,7 +1309,7 @@ static BOOL s_bHostWepRxEncryption (
     )
 {
     PSMgmtObject    pMgmt = &(pDevice->sMgmtObj);
-    UINT            PayloadLen = FrameSize;
+    unsigned int            PayloadLen = FrameSize;
     PBYTE           pbyIV;
     BYTE            byKeyIdx;
     BYTE            byDecMode = KEY_CTL_WEP;
@@ -1430,8 +1430,8 @@ static BOOL s_bHostWepRxEncryption (
 static BOOL s_bAPModeRxData (
      PSDevice pDevice,
      struct sk_buff *skb,
-     UINT     FrameSize,
-     UINT     cbHeaderOffset,
+     unsigned int     FrameSize,
+     unsigned int     cbHeaderOffset,
      signed int      iSANodeIndex,
      signed int      iDANodeIndex
     )
@@ -1504,7 +1504,8 @@ static BOOL s_bAPModeRxData (
             iDANodeIndex = 0;
 
         if ((pDevice->uAssocCount > 1) && (iDANodeIndex >= 0)) {
-            bRelayPacketSend(pDevice, (PBYTE)(skb->data + cbHeaderOffset), FrameSize, (UINT)iDANodeIndex);
+		bRelayPacketSend(pDevice, (PBYTE) (skb->data + cbHeaderOffset),
+				 FrameSize, (unsigned int) iDANodeIndex);
         }
 
         if (bRelayOnly)
